backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cdp-agentkit-core>=0.0.11",
    "flask>=3.1.0",
    "flask-cors>=5.0.0",
    "langchain>=0.3.17",
    "openai>=1.61.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "slither-analyzer>=0.10.3",
    "sseclient-py>=1.8.0",
]
